lib/protocols.c lib/rfc2068.c lib/rfc959.c lib/sshv2.c src/gtk/bookmarks.c
2004-7-27 Brian Masney <masneyb@gftp.org>

        * lib/protocols.c lib/rfc2068.c lib/rfc959.c lib/sshv2.c
        src/gtk/bookmarks.c src/gtk/dnd.c src/gtk/gftp-gtk.c src/gtk/gtui.c
        src/gtk/menu-items.c src/misc-gtk.c src/gtk/options_dialog.c
        src/gtk/transfer.c src/text/gftp-text.c src/uicommon/gftpui.c -
        various fixes found through a static code analysis. (some
        signed/unsigned fixes, removed unneeded casts, indentation fixes,
        added static declaration to some functions)

+++ lib/rfc2068.c
@@ -308,7 +308,7 @@ rfc2068_get_file (gftp_request * request, const char *filename, int fd,
 #if defined (_LARGEFILE_SOURCE)
       request->logging_function (gftp_logging_misc, request,
                               _("Starting the file transfer at offset %lld\n"),
-                              (long long) startsize);
+                              startsize);
 
       oldstr = tempstr;
       tempstr = g_strdup_printf ("%sRange: bytes=%lld-\n", tempstr, startsize);
@@ -607,7 +607,7 @@ rfc2068_get_next_file (gftp_request * request, gftp_file * fle, int fd)
 {
   rfc2068_params * params;
   char tempstr[8192];
-  ssize_t len;
+  size_t len;
   int ret;
 
   g_return_val_if_fail (request != NULL, GFTP_EFATAL);
